Split server functions to different file
